+++ src/io/midi_in.h
@@ -15,6 +15,7 @@
 #include "config/params.h"
 #include "config/synth_params.h"
 #include "config/gen_params.h"
+#include "config/gran_params.h"
 #include "mod/modulation.h"
 #include "fx/master.h"
 #include "io/knobs.h"
@@ -83,6 +84,9 @@ inline bool PumpMidi(daisy::MidiUsbHandler& midi, IMode* mode, ShiftKnobs& shift
         else if (n >= params::midi::kCcGenBase &&
                  n < params::midi::kCcGenBase + GP_COUNT)
           g_genParams.v[n - params::midi::kCcGenBase] = v;
+        else if (n >= params::midi::kCcGranBase &&
+                 n < params::midi::kCcGranBase + GR_COUNT)
+          g_granParams.v[n - params::midi::kCcGranBase] = v;
         else if (n == params::midi::kCcTempo)
           clock.SetInternalBpm(40.0f + v * 160.0f);            // 40..200 BPM
         else if (n == params::midi::kCcDelaySync)